backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dias Ziyada <dias.ziyada12345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
ruff = "^0.11.7"
mypy = "^1.15.0"
python-decouple = "^3.8"
dj-database-url = "^2.3.0"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
django-split-settings = "^1.3.2"
mixer = "^7.2.2"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.5.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
